# ComfyUI/openapi.yaml
components:
    schemas:
...
        ErrorResponse:
            description: Standard error response with a machine-readable code and human-readable message.
            properties:
                code:
                    type: string
                details:
                    additionalProperties: true
                    description: Optional open object carrying structured, machine-readable context about the error (e.g. offending field names, validation specifics). Absent for most errors; consumers must not assume any particular shape.
                    type: object
                message:
                    type: string
            required:
                - code
                - message
            type: object
...
        TaskResponse:
            description: Full task details including payload and result
            properties:
                completed_at:
                    description: When task completed or failed (null if not finished)
                    format: date-time
                    type: string
                create_time:
                    description: Task creation timestamp
                    format: date-time
                    type: string
                error_message:
                    description: Error message on failure (null if not failed)
                    type: string
                id:
                    description: Unique task identifier
                    format: uuid
                    type: string
                idempotency_key:
                    description: Caller-provided key for idempotent task creation
                    type: string
                payload:
                    additionalProperties: true
                    description: Task input data
                    type: object
                result:
                    additionalProperties: true
                    description: Task output data (null if not completed)
                    type: object
                started_at:
                    description: When task execution started (null if not started)
                    format: date-time
                    type: string
                status:
                    description: Current task status
                    enum:
                        - created
                        - running
                        - completed
                        - failed
                    type: string
                task_name:
                    description: Task type name (e.g., model_upload)
                    type: string
                update_time:
                    description: Task last update timestamp
                    format: date-time
                    type: string
            required:
                - id
                - idempotency_key
                - task_name
                - payload
                - status
                - create_time
                - update_time
            type: object
...
    securitySchemes:
        ApiKeyAuth:
            description: |
                API key authentication. Keys are prefixed with 'comfyui-' and can be
                generated from user account settings. Example: 'comfyui-abc123...'
            in: header
            name: X-API-Key
            type: apiKey
        BearerAuth:
            bearerFormat: JWT
            description: |
                Firebase JWT token authentication. Obtain a token by authenticating
                with Firebase and pass it in the Authorization header.
            scheme: bearer
            type: http
        CookieAuth:
            description: |
                Session cookie authentication. Set automatically after successful
                login via the /api/auth/session endpoint.
            in: cookie
            name: session
            type: apiKey
info:
    description: |
        API for ComfyUI - A powerful and modular UI for Stable Diffusion.

        This API allows you to interact with ComfyUI programmatically, including:
        - Retrieving prompt information
        - Retrieving node information
    license:
        name: GNU General Public License v3.0
        url: https://github.com/Comfy-Org/ComfyUI/blob/master/LICENSE
    title: ComfyUI API
    version: 1.0.0
openapi: 3.0.3
paths:
...
    /api/tasks/{task_id}:
        get:
            description: |
                Retrieve full details for a specific background task.
            operationId: getTask
            parameters:
                - description: Task identifier (UUID)
                  in: path
                  name: task_id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TaskResponse'
                    description: Success - Task details retrieved
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized - Authentication required
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Task not found (also returned for ownership failures to avoid leaking task existence)
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal server error
            summary: Get task details
            tags:
                - task
...
security:
    - ApiKeyAuth: []
    - BearerAuth: []
servers:
    - description: Default ComfyUI server
      url: /
...